+++ test/unit/services/RelayOrderService.test.ts
@@ -11,7 +11,7 @@ import { OnChainValidatorMap } from '../../../lib/handlers/OnChainValidatorMap'
 import { kickoffOrderTrackingSfn } from '../../../lib/handlers/shared/sfn'
 import { RelayOrder } from '../../../lib/models'
 import { BaseOrdersRepository } from '../../../lib/repositories/base'
-import { RelayOrderService } from '../../../lib/services/RelayOrderService/RelayOrderService'
+import { RelayOrderService } from '../../../lib/services/RelayOrderService'
 import { OffChainRelayOrderValidator } from '../../../lib/util/OffChainRelayOrderValidator'
 import { SDKRelayOrderFactory } from '../../factories/SDKRelayOrderFactory'
 import { QueryParamsBuilder } from '../builders/QueryParamsBuilder'
